+++ src/pages/player-detail/model/types.ts
@@ -1,9 +1,4 @@
-import type {
-  GrowthSummary,
-  PlayerDetailResponse,
-  Position,
-  SeasonHistory,
-} from '@/entities/player'
+import type { PlayerDetailResponse } from '@/entities/player'
 
 export type PlayerDetailTabLabel = '선수 히스토리' | '이적 예측'
 export type PlayerDetailTabSlug = 'history' | 'transfer'
@@ -21,31 +16,4 @@ export interface PlayerDetailDesktopHeaderProps {
 export interface PlayerDetailMobileHeaderProps {
   player: PlayerDetailResponse
   scrollProgress: number
-}
-
-export interface MarketValueChartProps {
-  history: SeasonHistory[]
-}
-
-export interface GrowthSummaryCardsProps {
-  summary: GrowthSummary
-  seasonCount: number
-}
-
-export interface SeasonStatsSectionProps {
-  history: SeasonHistory[]
-  position: Position
-}
-
-export interface SeasonStatsTableProps {
-  history: SeasonHistory[]
-}
-
-export interface SeasonStatsChartsProps {
-  history: SeasonHistory[]
-  position: Position
-}
-
-export interface PlayerHistoryTabProps {
-  position: Position
 }
