+++ src/widgets/history/ui/GrowthSummaryCards.tsx
