+++ src/widgets/history/ui/SeasonStatsTable.tsx
